修改配置和系统用户管理

+++ auth-server/src/main/java/com/ztw/iservice/IUserService.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by zsl-pc on 2016/8/31.
  */
-public interface IUserService extends JpaRepository<User, Long> {
+public interface IUserService extends JpaRepository<User, Integer> {
 
     public User findByUsername(String username);
 }
